% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_KPLS.R
\name{funcion_objetivo_2}
\alias{funcion_objetivo_2}
\title{Funcion Objetivo de Evaluacion Final}
\usage{
funcion_objetivo_2(x, datos1, datos2, numC)
}
\arguments{
\item{x}{valor de sigma.}

\item{datos1}{conjunto de datos de variables independientes.}

\item{datos2}{conjunto de datos de variables dependientes.}

\item{numC}{numero de componentes.}
}
\description{
Funcion Objetivo de Evaluacion Final
}
