% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_KPLS.R
\name{optimkpls}
\alias{optimkpls}
\title{Optimizacion de KPLS mediante un algoritmmo memetico}
\usage{
optimkpls(datos1, datos2, numC, Li, Ls, Pob, maxE)
}
\arguments{
\item{datos1}{conjunto de datos de variables independientes.}

\item{datos2}{conjunto de datos de variables dependientes.}

\item{numC}{numero de componentes}

\item{Li}{cota inferior}

\item{Ls}{cota superior}

\item{Pob}{numero de poblacion}

\item{maxE}{numero de evaluacion maxima}
}
\value{
sigma y numero de componente
}
\description{
Ajuste metaheurístico de parámetros de la regresión KPLS
}
\examples{
numCompoments <- 7
path<-"parkinson.xlsx"
dataset <- read_excel(path)
database <- data.frame(dataset)
data1<-database[2:20]
data2<-database[,21:22]
lo<-0
up<-100
po<-40
mE<-100
Resultado<-optim_kpls(data1,data2,numCompoments,lo,up,po,mE)
print(Resultado)
}
